Move DossierResolutionStatusmessageMixin to a separate file.
